move slice func under ccipcommon
