--- core/services/ocr2/plugins/ccip/internal/ccipcalc/slice.go
@@ -1,17 +1,0 @@
-package ccipcalc
-
-// FlattenUniqueSlice returns a flattened slice that contains unique elements.
-func FlattenUniqueSlice[T comparable](slices ...[]T) []T {
-	seen := make(map[T]struct{})
-	flattened := make([]T, 0)
-
-	for _, sl := range slices {
-		for _, el := range sl {
-			if _, exists := seen[el]; !exists {
-				flattened = append(flattened, el)
-				seen[el] = struct{}{}
-			}
-		}
-	}
-	return flattened
-}
