--- core/services/ocr2/plugins/ccip/internal/ccipcalc/slice_test.go
@@ -1,32 +1,0 @@
-package ccipcalc
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestFlattenUniqueSlice(t *testing.T) {
-	testCases := []struct {
-		name           string
-		inputSlices    [][]int
-		expectedOutput []int
-	}{
-		{name: "empty", inputSlices: nil, expectedOutput: []int{}},
-		{name: "empty 2", inputSlices: [][]int{}, expectedOutput: []int{}},
-		{name: "single", inputSlices: [][]int{{1, 2, 3, 3, 3, 4}}, expectedOutput: []int{1, 2, 3, 4}},
-		{name: "simple", inputSlices: [][]int{{1, 2, 3}, {2, 3, 4}}, expectedOutput: []int{1, 2, 3, 4}},
-		{
-			name:           "more complex case",
-			inputSlices:    [][]int{{1, 3}, {2, 4, 3}, {5, 2, -1, 7, 10}},
-			expectedOutput: []int{1, 3, 2, 4, 5, -1, 7, 10},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			res := FlattenUniqueSlice(tc.inputSlices...)
-			assert.Equal(t, tc.expectedOutput, res)
-		})
-	}
-}
