+++ core/services/ocr2/plugins/ccip/internal/ccipcommon/shortcuts.go
@@ -52,3 +52,19 @@ func GetDestinationTokens(ctx context.Context, offRamp ccipdata.OffRampReader, p
 
 	return destFeeTokens, destBridgeableTokens, nil
 }
+
+// FlattenUniqueSlice returns a flattened slice that contains unique elements.
+func FlattenUniqueSlice[T comparable](slices ...[]T) []T {
+	seen := make(map[T]struct{})
+	flattened := make([]T, 0)
+
+	for _, sl := range slices {
+		for _, el := range sl {
+			if _, exists := seen[el]; !exists {
+				flattened = append(flattened, el)
+				seen[el] = struct{}{}
+			}
+		}
+	}
+	return flattened
+}
